refact: split functions in files (add cmd_exec_redirects.c)
