+++ src/command/cmd_exec.c
@@ -12,132 +12,7 @@
 
 #include "minishell.h"
 
-static int	handle_heredoc(char *delimiter)
-{
-	int		pipe_fd[2];
-	char	*line;
-	size_t	delim_len;
-
-	if (pipe(pipe_fd) == -1)
-		return (-1);
-	delim_len = ft_strlen(delimiter);
-	while (1)
-	{
-		line = readline("pipe heredoc> ");
-		if (!line)
-		{
-			close(pipe_fd[1]);
-			return (-1);
-		}
-		if (ft_strlen(line) == delim_len && ft_strcmp(line, delimiter) == 0)
-		{
-			free(line);
-			break ;
-		}
-		write(pipe_fd[1], line, ft_strlen(line));
-		write(pipe_fd[1], "\n", 1);
-		free(line);
-	}
-	close(pipe_fd[1]);
-	return (pipe_fd[0]);
-}
-
-static int	setup_redirections(t_redirect *redirects)
-{
-	t_redirect	*current;
-	int			fd;
-	int			heredoc_fd;
-
-	heredoc_fd = -1;
-	current = redirects;
-	while (current)
-	{
-		if (current->op_type == OP_HEREDOC)
-		{
-			heredoc_fd = handle_heredoc(current->file);
-			if (heredoc_fd == -1)
-				return (-1);
-		}
-		current = current->next;
-	}
-	current = redirects;
-	while (current)
-	{
-		if (current->op_type == OP_REDIR_IN)
-		{
-			fd = open(current->file, O_RDONLY);
-			if (fd == -1)
-				return (-1);
-			dup2(fd, STDIN_FILENO);
-			close(fd);
-		}
-		else if (current->op_type == OP_HEREDOC)
-		{
-			dup2(heredoc_fd, STDIN_FILENO);
-			close(heredoc_fd);
-		}
-		else if (current->op_type == OP_REDIR_OUT)
-		{
-			fd = open(current->file, O_WRONLY | O_CREAT | O_TRUNC, 0644);
-			if (fd == -1)
-				return (-1);
-			dup2(fd, STDOUT_FILENO);
-			close(fd);
-		}
-		else if (current->op_type == OP_REDIR_APPEND)
-		{
-			fd = open(current->file, O_WRONLY | O_CREAT | O_APPEND, 0644);
-			if (fd == -1)
-				return (-1);
-			dup2(fd, STDOUT_FILENO);
-			close(fd);
-		}
-		current = current->next;
-	}
-	return (0);
-}
-
-void	cmd_exec_inline(int argc, char **argv, t_env *env, t_cmd *cmd)
-{
-	if (argc == 3 && argv[1] && ft_strncmp(argv[1], "-c", 3) == 0)
-	{
-		cmd_init(argv[2], cmd, env);
-		cmd_exec(cmd, env);
-		free_cmd(cmd);
-		env_free(env);
-		exit(g_signal);
-	}
-	else if (argc > 1)
-	{
-		printf("Usage:\n./minishell\nOR\n./minishell -c \"command\"\n");
-		g_signal = 2;
-	}
-}
-
-static void handle_redirections(t_cmdblock *block, int *pipefd, int *fd_output)
-{
-	t_redirect *redir;
-
-	*fd_output = STDOUT_FILENO;
-	if (block->redirects)
-	{
-		if (block->next)
-			close(pipefd[1]);
-		redir = block->redirects;
-		while (redir)
-		{
-			if (redir->op_type == OP_REDIR_OUT)
-				*fd_output = open(redir->file, O_WRONLY | O_CREAT | O_TRUNC, 0644);
-			else if (redir->op_type == OP_REDIR_APPEND)
-				*fd_output = open(redir->file, O_WRONLY | O_CREAT | O_APPEND, 0644);
-			redir = redir->next;
-		}
-	}
-	else if (block->next)
-		*fd_output = pipefd[1];
-}
-
-static void handle_pipe_setup(t_cmdblock *block, int *pipefd, int *prev_pipe)
+static void	handle_pipe_setup(t_cmdblock *block, int *pipefd, int *prev_pipe)
 {
 	if (*prev_pipe != STDIN_FILENO)
 		close(*prev_pipe);
@@ -157,7 +32,8 @@ static void	exec_piped_cmd(t_cmd *cmd, t_env *env,
 
 	if (cmd_setup(cmd, env, &args, &full_path) != 0)
 		exit(g_signal);
-	if (cmd->cmd->redirects && setup_redirections(cmd->cmd->redirects) == -1)
+	if (cmd->cmd->redirects
+		&& cmd_exec_setup_redirect(cmd->cmd->redirects) == -1)
 	{
 		perror("redirect error");
 		exit(1);
@@ -207,7 +83,7 @@ int	cmd_exec(t_cmd *cmd, t_env *env)
 		}
 		if (get_builtin(cmdtmp->cmd->exec))
 		{
-			handle_redirections(cmdtmp->cmd, pipefd, &fd_output);
+			cmd_exec_handle_redirect(cmdtmp->cmd, pipefd, &fd_output);
 			result = execute_builtin(cmdtmp, env, prev_pipe, fd_output);
 			if (fd_output != STDOUT_FILENO && fd_output != pipefd[1])
 				close(fd_output);
@@ -228,9 +104,7 @@ int	cmd_exec(t_cmd *cmd, t_env *env)
 					close(pipefd[0]);
 				}
 				else
-				{
 					exec_piped_cmd(cmdtmp, env, prev_pipe, STDOUT_FILENO);
-				}
 			}
 		}
 		handle_pipe_setup(cmdtmp->cmd, pipefd, &prev_pipe);
